+++ HISTORY.md
@@ -68,6 +68,8 @@
   statements.
 - Added support for VHDL-2008 extended bit string literals.
 - Non-globally-static actuals allowed in port maps in VHDL-2008 mode.
+- Added support for VHDL-2008 sequential conditional signal assignment
+  statements.
 
 ## 1.5 - 2020-07-19
 - IEEE library sources are now distributed